Refactor DiskPaper and rename to DiskGraph
